La til mulighet for brukere å slette egne saker på Min side, flyttet Mypage fra Home til Bruker mappe og BrukerController.
