+++ kartverketprosjekt/Views/Bruker/Mypage.cshtml
@@ -1,0 +1,112 @@
+@model List<kartverketprosjekt.Models.SakModel>
+
+@{
+    Layout = "~/Views/Shared/_Layout.cshtml";
+    ViewData["Title"] = "Min Side";
+}
+
+<a asp-controller="Home" asp-action="Index">
+    <i class='bx bx-chevron-left'></i>
+</a>
+<h1>Min Side</h1>
+
+
+@if (TempData["SuccessMessage"] != null)
+{
+    <div class="alert alert-success">
+        @TempData["SuccessMessage"]
+    </div>
+}
+
+@if (TempData["ErrorMessage"] != null)
+{
+    <div class="alert alert-danger">
+        @TempData["ErrorMessage"]
+    </div>
+}
+
+<div class="dashboard">
+    <div class="dashboard-box">
+ <h2>Kontoinstillinger </h2>
+        <!-- Form for Updating Name -->
+        <form asp-controller="Bruker" asp-action="UpdateName" method="post">
+            @Html.AntiForgeryToken() <!-- Ensure this is included for security -->
+                <label for="Navn">Oppdater navnet ditt</label>
+                <input type="text" id="Navn" name="navn" placeholder="Skriv inn nytt navn" required />
+            <button type="submit" class="btn-primary">Oppdater Navn</button>
+        </form>
+        <!-- Endre Passord Form -->
+        <form asp-controller="Bruker" asp-action="UpdatePassword" method="post">
+            @Html.AntiForgeryToken()
+                <label for="Password">Nytt passord:</label>
+                <input type="password" id="Password" name="Password" placeholder="Skriv inn nytt passord" required />
+            <button type="submit" class="btn-primary">Oppdater Passord</button>
+        </form>
+
+    </div>
+    <style>
+        .status-ikke-påbegynt {
+            color: grey;
+        }
+
+        .status-in-progress {
+            color: orange;
+        }
+
+        .status-completed {
+            color: Green;
+        }
+
+        .status-closed {
+            color: red;
+        }
+
+        .status-arkivert {
+            color: Brown;
+        }
+    </style>
+    <div class="table-container">
+        <div class="dashboard-box">
+            <h2>Oversikt over dine saker</h2>
+            @if (Model != null && Model.Count > 0)
+            {
+                <ul>
+                    @foreach (var sak in Model)
+                    {
+                        // Assign the appropriate CSS class based on the status
+                        string statusClass = sak.status switch
+                        {
+                            "Ubehandlet" => "status-ikke-påbegynt",
+                            "Under Behandling" => "status-in-progress",
+                            "Løst" => "status-completed",
+                            "Avvist" => "status-closed",
+                            "Arkivert" => "status-arkivert", // Added case for 'Arkivert'
+                            _ => string.Empty
+                        };
+
+                        <li style="list-style-type: none;">
+                            <strong>@sak.Kommunenavn</strong> -
+                            <strong>Status: <span class="@statusClass">@sak.status</span></strong> -
+                            Opprettet: @sak.opprettet_dato.ToString("d")
+
+                            <form class="hidden-form" asp-controller="Bruker" asp-action="DeleteCase" method="post">
+                                <input type="hidden" name="id" value="@sak.id" />
+                                <button style="all: unset; cursor: pointer;" type="submit" onclick="return confirm('Er du sikker på at du vil slette denne saken?');">
+                                    <i class='bx bx-x' style="font-size: 30px; color: var(--text-color);"></i>
+                                </button>
+                            </form>
+                        </li>
+                    }
+                </ul>
+            }
+            else
+            {
+                <p>Du har ingen saker.</p>
+            }
+        </div>
+    </div>
+
+</div>
+@{
+    ViewData["RenderFooter"] = true;
+}
