+++ kartverketprosjekt/Views/Saksbehandler/CaseWorkerView.cshtml
@@ -225,7 +225,7 @@
                     </ul>
                 </div>
                 <div class="btn-container">
-                    <button class="btn btn-danger" id="slettSakKnapp">Slett sak</button>
+                    <button class="btn btn-danger" style="margin-top: 70px;" id="slettSakKnapp">Slett sak</button>
             <!--    <button class="btn-primary" id="lagreKnapp">Lagre endringer</button>   Denne knappen har ingen funksjon. -->
                 </div>
             </div>
